+++ packages/jest-reporters/src/GitHubActionsReporter.ts
@@ -5,19 +5,17 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {
-  summary as actionsSummary,
-  error as errorAnnotation,
-} from '@actions/core';
+import {summary as actionsSummary} from '@actions/core';
 import stripAnsi = require('strip-ansi');
 import type {
   AggregatedResult,
   Test,
-  TestCaseResult,
   TestContext,
+  TestResult,
 } from '@jest/test-result';
+import type {Config} from '@jest/types';
 import {
-  // formatPath,
+  formatPath,
   getStackTraceLines,
   getTopFrame,
   separateMessageFromStack,
@@ -25,32 +23,39 @@ import {
 import {pluralize} from 'jest-util';
 import BaseReporter from './BaseReporter';
 
-const errorTitleSeparator = ' \u203A ';
+type AnnotationOptions = {
+  file?: string;
+  line?: number | string;
+  message: string;
+  title: string;
+  type: 'error' | 'warning';
+};
+
+const titleSeparator = ' \u203A ';
 
 export default class GitHubActionsReporter extends BaseReporter {
   static readonly filename = __filename;
 
-  override onTestCaseResult(
-    test: Test,
-    {failureMessages, ancestorTitles, title}: TestCaseResult,
-  ): void {
-    failureMessages.forEach(failureMessage => {
-      const {message, stack} = separateMessageFromStack(failureMessage);
+  onTestFileResult({context}: Test, {testResults}: TestResult): void {
+    testResults.forEach(result => {
+      const title = [...result.ancestorTitles, result.title].join(
+        titleSeparator,
+      );
 
-      const stackLines = getStackTraceLines(stack);
-      // const formattedLines = stackLines.map(line =>
-      //   formatPath(line, test.context.config, null),
-      // );
-      const topFrame = getTopFrame(stackLines);
+      result.retryReasons?.forEach((retryReason, index) => {
+        this.#createAnnotation({
+          ...this.#getMessageDetails(retryReason, context.config),
+          title: `RETRY ${index + 1}: ${title}`,
+          type: 'warning',
+        });
+      });
 
-      const errorTitle = [...ancestorTitles, title].join(errorTitleSeparator);
-      const errorMessage = stripAnsi([message, ...stackLines].join('\n'));
-
-      errorAnnotation(errorMessage, {
-        file: test.path,
-        startColumn: topFrame?.column,
-        startLine: topFrame?.line,
-        title: errorTitle,
+      result.failureMessages.forEach(failureMessage => {
+        this.#createAnnotation({
+          ...this.#getMessageDetails(failureMessage, context.config),
+          title,
+          type: 'error',
+        });
       });
     });
   }
@@ -137,5 +142,34 @@ export default class GitHubActionsReporter extends BaseReporter {
     ]);
 
     await summary.write();
+  }
+
+  #getMessageDetails(failureMessage: string, config: Config.ProjectConfig) {
+    const {message, stack} = separateMessageFromStack(failureMessage);
+
+    const stackLines = getStackTraceLines(stack);
+    const topFrame = getTopFrame(stackLines);
+
+    const normalizedStackLines = stackLines.map(line =>
+      formatPath(line, config),
+    );
+    const messageText = [message, ...normalizedStackLines].join('\n');
+
+    return {
+      file: topFrame?.file,
+      line: topFrame?.line,
+      message: messageText,
+    };
+  }
+
+  #createAnnotation({file, line, message, title, type}: AnnotationOptions) {
+    message = stripAnsi(
+      // copied from: https://github.com/actions/toolkit/blob/main/packages/core/src/command.ts
+      message.replace(/%/g, '%25').replace(/\r/g, '%0D').replace(/\n/g, '%0A'),
+    );
+
+    this.log(
+      `\n::${type} file=${file},line=${line},title=${title}::${message}`,
+    );
   }
 }