+++ packages/jest-reporters/src/__tests__/GitHubActionsReporter.test.ts
@@ -5,17 +5,17 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {
-  summary as actionsSummary,
-  error as errorAnnotation,
-} from '@actions/core';
+import {summary as actionsSummary} from '@actions/core';
 import type {
   AggregatedResult,
   Test,
   TestCaseResult,
   TestContext,
+  TestResult,
 } from '@jest/test-result';
 import GitHubActionsReporter from '../GitHubActionsReporter';
+
+process.stderr.write = jest.fn();
 
 jest.spyOn(Date, 'now').mockReturnValue(6000);
 
@@ -30,7 +30,7 @@ jest.mock('@actions/core', () => {
     write,
   } as unknown as jest.Mocked<typeof actionsSummary>;
 
-  return {error: jest.fn(), summary};
+  return {summary};
 });
 
 afterEach(() => {
@@ -41,7 +41,6 @@ const reporter = new GitHubActionsReporter();
 
 const testMeta = {
   context: {config: {rootDir: '/user/project'}},
-  path: '/user/project/__tests__/quick.test.js',
 } as Test;
 
 const expectationsErrorMessage =
@@ -49,7 +48,7 @@ const expectationsErrorMessage =
   '\n' +
   'Expected: \x1B[32m1\x1B[39m\n' +
   'Received: \x1B[31m10\x1B[39m\n' +
-  '    at Object.toBe (/user/project/__tests__/quick.test.js:20:14)\n' +
+  '    at Object.toBe (/user/project/__tests__/example.test.js:20:14)\n' +
   '    at Promise.then.completed (/user/project/jest/packages/jest-circus/build/utils.js:333:28)\n' +
   '    at new Promise (<anonymous>)\n' +
   '    at callAsyncCircusFn (/user/project/jest/packages/jest-circus/build/utils.js:259:10)\n' +
@@ -62,7 +61,7 @@ const expectationsErrorMessage =
 
 const referenceErrorMessage =
   'ReferenceError: abc is not defined\n' +
-  '    at Object.abc (/user/project/__tests__/quick.test.js:25:12)\n' +
+  '    at Object.abc (/user/project/__tests__/example.test.js:25:12)\n' +
   '    at Promise.then.completed (/user/project/jest/packages/jest-circus/build/utils.js:333:28)\n' +
   '    at new Promise (<anonymous>)\n' +
   '    at callAsyncCircusFn (/user/project/jest/packages/jest-circus/build/utils.js:259:10)\n' +
@@ -73,82 +72,90 @@ const referenceErrorMessage =
   '    at _runTestsForDescribeBlock (/user/project/jest/packages/jest-circus/build/run.js:90:9)\n' +
   '    at run (/user/project/jest/packages/jest-circus/build/run.js:31:3)';
 
+const retryErrorMessage =
+  'Error: \x1B[2mexpect(\x1B[22m\x1B[31mreceived\x1B[39m\x1B[2m).\x1B[22mtoBeFalsy\x1B[2m()\x1B[22m\n' +
+  '\n' +
+  'Received: \x1B[31mtrue\x1B[39m\n' +
+  '    at Object.toBeFalsy (/user/project/__tests__/example.test.js:19:20)\n' +
+  '    at Promise.then.completed (/user/project/jest/packages/jest-circus/build/utils.js:333:28)\n' +
+  '    at new Promise (<anonymous>)\n' +
+  '    at callAsyncCircusFn (/user/project/jest/packages/jest-circus/build/utils.js:259:10)\n' +
+  '    at _callCircusTest (/user/project/jest/packages/jest-circus/build/run.js:276:40)\n' +
+  '    at _runTest (/user/project/jest/packages/jest-circus/build/run.js:208:3)\n' +
+  '    at _runTestsForDescribeBlock (/user/project/jest/packages/jest-circus/build/run.js:96:9)\n' +
+  '    at _runTestsForDescribeBlock (/user/project/jest/packages/jest-circus/build/run.js:90:9)\n' +
+  '    at run (/user/project/jest/packages/jest-circus/build/run.js:31:3)\n' +
+  '    at runAndTransformResultsToJestFormat (/user/project/jest/packages/jest-circus/build/legacy-code-todo-rewrite/jestAdapterInit.js:135:21)';
+
 const testCaseResult = {
   ancestorTitles: [] as Array<string>,
   failureMessages: [expectationsErrorMessage],
-  title: 'some test',
+  title: 'example test',
 } as TestCaseResult;
 
 const testContexts = new Set<TestContext>();
 
-describe("passes test case report to '@actions/core'", () => {
-  test('when expect returns an error', () => {
-    reporter.onTestCaseResult(testMeta, {
-      ...testCaseResult,
-      failureMessages: [expectationsErrorMessage],
-    });
+describe('logs error annotation', () => {
+  test('when an expectation fails to pass', () => {
+    reporter.onTestFileResult(testMeta, {
+      testResults: [
+        {
+          ...testCaseResult,
+          failureMessages: [expectationsErrorMessage],
+        },
+      ],
+    } as TestResult);
 
-    const expectedMessage =
-      'expect(received).toBe(expected) // Object.is equality\n' +
-      '\n' +
-      'Expected: 1\n' +
-      'Received: 10\n' +
-      '\n' +
-      '    at Object.toBe (__tests__/quick.test.js:20:14)';
-
-    expect(errorAnnotation).toBeCalledWith(expectedMessage, {
-      file: expect.any(String),
-      startColumn: 14,
-      startLine: 20,
-      title: expect.any(String),
-    });
+    expect(jest.mocked(process.stderr.write)).toBeCalledTimes(1);
+    expect(jest.mocked(process.stderr.write).mock.calls[0]).toMatchSnapshot();
   });
 
   test('when a test has reference error', () => {
-    reporter.onTestCaseResult(
-      {...testMeta, path: '/user/project/__tests__/quick.test.js:25:12'},
+    reporter.onTestFileResult(
+      {...testMeta, path: '/user/project/__tests__/example.test.js:25:12'},
       {
-        ...testCaseResult,
-        failureMessages: [referenceErrorMessage],
-      },
+        testResults: [
+          {
+            ...testCaseResult,
+            failureMessages: [referenceErrorMessage],
+          },
+        ],
+      } as TestResult,
     );
 
-    const expectedMessage =
-      'ReferenceError: abc is not defined\n' +
-      '\n' +
-      '    at Object.abc ([broken]__tests__/quick.test.js:25:12)';
-
-    expect(errorAnnotation).toBeCalledWith(expectedMessage, {
-      file: expect.any(String),
-      startColumn: 12,
-      startLine: 25,
-      title: expect.any(String),
-    });
+    expect(jest.mocked(process.stderr.write)).toBeCalledTimes(1);
+    expect(jest.mocked(process.stderr.write).mock.calls[0]).toMatchSnapshot();
   });
 
   test('when test is wrapped in describe block', () => {
-    reporter.onTestCaseResult(testMeta, {
-      ...testCaseResult,
-      ancestorTitles: ['describe'],
-    });
+    reporter.onTestFileResult(testMeta, {
+      testResults: [
+        {
+          ...testCaseResult,
+          ancestorTitles: ['describe'],
+        },
+      ],
+    } as TestResult);
 
-    expect(errorAnnotation).toBeCalledWith(expect.any(String), {
-      file: '/user/project/__tests__/quick.test.js',
-      startColumn: 14,
-      startLine: 20,
-      title: 'describe \u203A some test',
-    });
+    expect(jest.mocked(process.stderr.write)).toBeCalledTimes(1);
+    expect(jest.mocked(process.stderr.write).mock.calls[0]).toMatchSnapshot();
   });
+});
 
-  test('when test not is wrapped in describe block', () => {
-    reporter.onTestCaseResult(testMeta, testCaseResult);
+describe('logs warning annotation before logging errors', () => {
+  test('when test result includes retry reasons', () => {
+    reporter.onTestFileResult(testMeta, {
+      testResults: [
+        {
+          ...testCaseResult,
+          failureMessages: [retryErrorMessage],
+          retryReasons: [retryErrorMessage],
+        },
+      ],
+    } as TestResult);
 
-    expect(errorAnnotation).toBeCalledWith(expect.any(String), {
-      file: '/user/project/__tests__/quick.test.js',
-      startColumn: 14,
-      startLine: 20,
-      title: 'some test',
-    });
+    expect(jest.mocked(process.stderr.write)).toBeCalledTimes(2);
+    expect(jest.mocked(process.stderr.write).mock.calls).toMatchSnapshot();
   });
 });