+++ scripts/build.mjs
@@ -62,8 +62,9 @@ function getBuildPath(file, buildFolder) {
 
 function buildNodePackage({packageDir, pkg}) {
   const srcDir = path.resolve(packageDir, SRC_DIR);
-  const pattern = path.resolve(srcDir, '**/*');
-  const files = glob.sync(pattern, {nodir: true});
+  const files = glob
+    .sync('**/*', {cwd: srcDir, nodir: true})
+    .map(res => path.join(srcDir, res));
 
   process.stdout.write(adjustToTerminalWidth(`${pkg.name}\n`));